moved overlay parser to global level
