+++ sqlglot/dialects/databricks.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from copy import deepcopy
 from collections import defaultdict
-import typing as t
 
 from sqlglot import exp, transforms, jsonpath, parser
 from sqlglot.dialects.dialect import (
@@ -75,20 +74,6 @@ class Databricks(Spark):
                 to=to,
             ),
         }
-
-        def _parse_overlay(self) -> exp.Overlay:
-            def _parse_overlay_arg(text: str) -> t.Optional[exp.Expression]:
-                return (
-                    self._match(TokenType.COMMA) or self._match_text_seq(text)
-                ) and self._parse_bitwise()
-
-            return self.expression(
-                exp.Overlay,
-                this=self._parse_bitwise(),
-                expression=_parse_overlay_arg("PLACING"),
-                from_=_parse_overlay_arg("FROM"),
-                for_=_parse_overlay_arg("FOR"),
-            )
 
     class Generator(Spark.Generator):
         TABLESAMPLE_SEED_KEYWORD = "REPEATABLE"
